requirements.txt
livekit-agents==1.2.15
livekit-agents[openai]
livekit-agents[silero]
python-dotenv